Added validation to edit and delete req-s

+++ Application/Recipes/Delete.cs
@@ -1,3 +1,4 @@
+using Application.Core;
 using MediatR;
 using Persistence;
 
@@ -5,12 +6,12 @@ namespace Application.Recipes
 {
     public class Delete
     {
-        public class Command : IRequest
+        public class Command : IRequest<Result<Unit>>
         {
             public Guid Id { get; set; }
         }
 
-        public class Handler : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command, Result<Unit>>
         {
             private readonly DataContext _context;
             public Handler(DataContext context)
@@ -18,14 +19,25 @@ namespace Application.Recipes
                 _context = context;
             }
 
-            public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var recipe = await _context.Recipes.FindAsync(request.Id);
+                
+                if (recipe == null) 
+                {
+                    return null;
+                }
+                
                 _context.Recipes.Remove(recipe);
 
-                await _context.SaveChangesAsync();
+                var result = await _context.SaveChangesAsync() > 0;
 
-                return Unit.Value;
+                if (!result)
+                {
+                    return Result<Unit>.Failure("Failed to delete the recipe");
+                }
+
+                return Result<Unit>.Success(Unit.Value);
             }
         }
     }
